+++ ibug/face_pose_augmentation/fpa/cpp/MM3D.h
@@ -2,10 +2,10 @@
 #define MM3D_H
 
 namespace MM3D {
-	void ZBuffer(double* vertex, double* tri, double* texture, int nver, int ntri,
-	    double* src_img, int width, int height, int nChannels, double* img, double* tri_ind);
-	void ZBufferTri(double* vertex, double* tri, double* texture_tri, int nver, int ntri,
-	    double* src_img, int width, int height, int nChannels, double* img, double* tri_ind);
+	void ZBuffer(double* vertex, int* tri, double* texture, int nver, int ntri,
+	    double* src_img, int width, int height, int nChannels, double* img, int* tri_ind);
+	void ZBufferTri(double* vertex, int* tri, double* texture_tri, int nver, int ntri,
+	    double* src_img, int width, int height, int nChannels, double* img, int* tri_ind);
 	bool PointInTri(double point[], double pt1[], double pt2[], double pt3[]);
 };
